# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: HIOS API
  version: 0.3.0
  description: |
    Sprint 1 implementation contract. RS256 bearer JWT identifies one tenant;
    active database membership is required. Paths remain unversioned.
servers:
  - url: https://api.example.invalid
    description: Placeholder API endpoint
paths:
  /health:
    get:
      summary: Process liveness
      description: Reports liveness only; does not certify dependency readiness.
      operationId: getHealth
      tags: [system]
      security: []
      responses:
        '200':
          description: Process is alive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'
              example: {status: ok}
  /plants:
    get:
      summary: List solar plants in the authorized tenant
      description: |
        Resolve exactly one authorized tenant from authenticated context before
        filtering and pagination. Missing or forbidden tenant context yields 403.
        tenant_id is a signed JWT claim and must match active database membership.
        Sort by stable public plant ID ascending. Offset pagination does not
        guarantee a snapshot while plants are added or removed.
      operationId: listPlants
      tags: [plants]
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          description: Maximum number of plants returned.
          schema: {type: integer, minimum: 1, maximum: 100, default: 50}
        - name: offset
          in: query
          description: Number of tenant-filtered plants to skip.
          schema: {type: integer, minimum: 0, maximum: 2147483647, default: 0}
      responses:
        '200':
          description: Tenant-filtered page; data is empty when no records remain.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlantList'
              example:
                data:
                  - {id: plant-001, name: Solar One, capacityKw: 120.5}
                limit: 50
                offset: 0
        '400':
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: {code: VALIDATION_ERROR, message: Invalid pagination parameters, requestId: req-001}
        '401':
          description: Missing or invalid authentication
          headers:
            WWW-Authenticate:
              required: true
              schema: {type: string}
              example: Bearer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: {code: AUTH_REQUIRED, message: Authentication required, requestId: req-002}
        '403':
          description: Tenant context unresolved or access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: {code: ACCESS_DENIED, message: Access denied, requestId: req-003}
        '500':
          description: Unexpected failure; no internal details exposed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: {code: INTERNAL_ERROR, message: Internal error, requestId: req-004}
    post:
      summary: Create a tenant-scoped solar plant and its manual passport
      description: |
        Stores only operator-entered engineering data. Credentials, tokens and
        remote-control settings are rejected by design; cloud access is added
        separately as a metadata-only read-only binding.
      operationId: createPlant
      tags: [plants]
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name]
              properties:
                name: {type: string, minLength: 1}
                capacityKw: {type: number, minimum: 0}
                latitude: {type: number, minimum: -90, maximum: 90}
                longitude: {type: number, minimum: -180, maximum: 180}
                timezone: {type: string, minLength: 1}
                capacityAcKw: {type: number, minimum: 0}
                tiltDeg: {type: number, minimum: 0, maximum: 90}
                azimuthDeg: {type: number, minimum: 0, exclusiveMaximum: 360}
                mountingType: {type: string, minLength: 1}
                meterBoundary: {type: string, minLength: 1}
                commissioningDate: {type: string, format: date}
                operatorNotes: {type: string, minLength: 1}
      responses:
        '201': {description: Plant created}
        '400': {description: Invalid plant passport}
        '401': {description: Authentication required}
        '403': {description: Tenant membership is inactive}
  /plants/{plant_id}/forecast-runs:
    get:
      summary: List immutable forecast runs for one authorized plant
      description: |
        Returns a newest-first page only when the plant belongs to the signed
        tenant and the caller has active database membership. This is a
        read-only index; it does not start, re-run or alter a forecast.
      operationId: listPlantForecastRuns
      tags: [forecasting]
      security:
        - bearerAuth: []
      parameters:
        - name: plant_id
          in: path
          required: true
          schema: {type: string, minLength: 1}
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 100, default: 50}
        - name: offset
          in: query
          schema: {type: integer, minimum: 0, maximum: 2147483647, default: 0}
      responses:
        '200':
          description: Tenant-authorized newest-first page.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ForecastRunList'}
              example:
                data:
                  - id: 00000000-0000-0000-0000-000000000001
                    forecastOriginUtc: '2026-09-11T00:00:00+00:00'
                    horizonId: day_ahead
                    modelId: MODEL-001
                    modelVersion: 1.0.0
                    featureVersion: model-001-features-v1
                    status: normal
                    pointCount: 24
                limit: 50
                offset: 0
        '400':
          description: Invalid pagination parameters.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
              example:
                error: {code: VALIDATION_ERROR, message: Invalid pagination parameters, requestId: req-010}
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
              example:
                error: {code: AUTH_REQUIRED, message: Authentication required, requestId: req-011}
        '403':
          description: Tenant context unresolved or membership inactive.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
              example:
                error: {code: ACCESS_DENIED, message: Access denied, requestId: req-012}
        '404':
          description: Plant is not visible in the authorized tenant.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
              example:
                error: {code: NOT_FOUND, message: Not found, requestId: req-013}
        '500':
          description: Unexpected failure; no internal details exposed.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
              example:
                error: {code: INTERNAL_ERROR, message: Internal error, requestId: req-014}
  /plants/{plant_id}/onboarding:
    get:
      summary: Read the manual passport and metadata-only cloud bindings for one authorized plant
      operationId: getPlantOnboarding
      tags: [plants, integrations]
      security: [{bearerAuth: []}]
      parameters:
        - {name: plant_id, in: path, required: true, schema: {type: string, minLength: 1}}
      responses:
        '200': {description: Tenant-authorized onboarding record}
        '401': {description: Authentication required}
        '404': {description: Plant not visible to this tenant}
  /plants/{plant_id}/cloud-bindings:
    post:
      summary: Register a read-only cloud binding using secret and consent references
      operationId: createPlantCloudBinding
      tags: [integrations]
      security: [{bearerAuth: []}]
      parameters:
        - {name: plant_id, in: path, required: true, schema: {type: string, minLength: 1}}
      responses:
        '201': {description: Read-only binding registered}
        '400': {description: Invalid connector metadata}
        '401': {description: Authentication required}
        '404': {description: Plant not visible to this tenant}
  /forecast-runs/{run_id}:
    get:
      summary: Read one immutable forecast run and a page of forecast points
      description: |
        Returns only a run owned by the tenant in the signed JWT and active
        database membership. The endpoint is read-only: it never invokes a
        provider, starts a job, changes a model or republishes a forecast.
      operationId: getForecastRun
      tags: [forecasting]
      security:
        - bearerAuth: []
      parameters:
        - name: run_id
          in: path
          required: true
          schema: {type: string, format: uuid}
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 100, default: 50}
        - name: offset
          in: query
          schema: {type: integer, minimum: 0, maximum: 2147483647, default: 0}
      responses:
        '200':
          description: Tenant-authorized immutable run and point page.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ForecastRunPage'}
              example:
                data:
                  id: 00000000-0000-0000-0000-000000000001
                  plantId: plant-001
                  forecastOriginUtc: '2026-09-11T00:00:00+00:00'
                  horizonId: day_ahead
                  modelId: MODEL-001
                  modelVersion: 1.0.0
                  featureVersion: model-001-features-v1
                  status: normal
                  points:
                    - intervalStartUtc: '2026-09-11T01:00:00+00:00'
                      intervalEndUtc: '2026-09-11T02:00:00+00:00'
                      predictedPowerKw: 12.5
                      predictedEnergyKwh: 12.5
                      qualityFlags: []
                      providerProvenance: {irradiance_global_wm2: solcast}
                limit: 50
                offset: 0
        '400':
          description: Invalid run ID or pagination parameters.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
              example:
                error: {code: VALIDATION_ERROR, message: Invalid pagination parameters, requestId: req-005}
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
              example:
                error: {code: AUTH_REQUIRED, message: Authentication required, requestId: req-006}
        '403':
          description: Tenant context unresolved or membership inactive.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
              example:
                error: {code: ACCESS_DENIED, message: Access denied, requestId: req-007}
        '404':
          description: No run is visible in the authorized tenant.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
              example:
                error: {code: NOT_FOUND, message: Not found, requestId: req-008}
        '500':
          description: Unexpected failure; no internal details exposed.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
              example:
                error: {code: INTERNAL_ERROR, message: Internal error, requestId: req-009}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 JWT with required exp, iat, iss, aud and sub; signed tenant_id plus active database membership. See ADR-0001.
  schemas:
    Health:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status: {type: string, enum: [ok]}
    Plant:
      type: object
      additionalProperties: false
      required: [id, name]
      properties:
        id:
          type: string
          minLength: 1
          description: Stable opaque public identifier.
        name: {type: string, minLength: 1}
        capacityKw:
          type: number
          minimum: 0
          description: Installed DC nameplate capacity in kW (kWp); omitted when unknown.
    PlantList:
      type: object
      additionalProperties: false
      required: [data, limit, offset]
      properties:
        data:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/Plant'
        limit: {type: integer, minimum: 1, maximum: 100}
        offset: {type: integer, minimum: 0}
    ForecastPoint:
      type: object
      additionalProperties: false
      required: [intervalStartUtc, intervalEndUtc, predictedPowerKw, predictedEnergyKwh, qualityFlags, providerProvenance]
      properties:
        intervalStartUtc: {type: string, format: date-time}
        intervalEndUtc: {type: string, format: date-time}
        predictedPowerKw: {type: number, minimum: 0}
        predictedEnergyKwh: {type: number, minimum: 0}
        qualityFlags: {type: array, items: {type: string}}
        providerProvenance: {type: object, additionalProperties: {type: string}}
    ForecastRun:
      type: object
      additionalProperties: false
      required: [id, plantId, forecastOriginUtc, horizonId, modelId, modelVersion, featureVersion, status, points]
      properties:
        id: {type: string, format: uuid}
        plantId: {type: string, minLength: 1}
        forecastOriginUtc: {type: string, format: date-time}
        horizonId: {type: string, enum: [intraday, day_ahead]}
        modelId: {type: string, minLength: 1}
        modelVersion: {type: string, minLength: 1}
        featureVersion: {type: string, minLength: 1}
        status: {type: string, enum: [normal, degraded, blocked]}
        points: {type: array, maxItems: 100, items: {$ref: '#/components/schemas/ForecastPoint'}}
    ForecastRunSummary:
      type: object
      additionalProperties: false
      required: [id, forecastOriginUtc, horizonId, modelId, modelVersion, featureVersion, status, pointCount]
      properties:
        id: {type: string, format: uuid}
        forecastOriginUtc: {type: string, format: date-time}
        horizonId: {type: string, enum: [intraday, day_ahead]}
        modelId: {type: string, minLength: 1}
        modelVersion: {type: string, minLength: 1}
        featureVersion: {type: string, minLength: 1}
        status: {type: string, enum: [normal, degraded, blocked]}
        pointCount: {type: integer, minimum: 0}
    ForecastRunList:
      type: object
      additionalProperties: false
      required: [data, limit, offset]
      properties:
        data: {type: array, maxItems: 100, items: {$ref: '#/components/schemas/ForecastRunSummary'}}
        limit: {type: integer, minimum: 1, maximum: 100}
        offset: {type: integer, minimum: 0}
    ForecastRunPage:
      type: object
      additionalProperties: false
      required: [data, limit, offset]
      properties:
        data: {$ref: '#/components/schemas/ForecastRun'}
        limit: {type: integer, minimum: 1, maximum: 100}
        offset: {type: integer, minimum: 0}
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId]
          properties:
            code: {type: string, minLength: 1}
            message: {type: string, minLength: 1}
            requestId: {type: string, minLength: 1}
            details:
              type: object
              description: Optional safe context; never include secrets or stack traces.